pyproject.toml
[tool.poetry]
name = "cancaonova-practice-schedule"
version = "0.1.0"
description = ""
authors = ["yudaihomma <kasettoc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.88.0"
google-auth-httplib2 = "^0.1.0"
google-auth = "^2.19.1"
tweepy = "^4.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
